Moved callbacks in own function
